+++ src/main/java/site/nomoreparties/stellarburgers/support/driver/DriverSupportManager.java
@@ -1,0 +1,2 @@
+package site.nomoreparties.stellarburgers.support.driver;public class DriverSupportManager {
+}
