--- src/test/java/site/nomoreparties/stellarburgers/ClassTest.java
@@ -1,18 +1,0 @@
-package site.nomoreparties.stellarburgers;
-
-import org.junit.After;
-import org.junit.Test;
-
-public class ClassTest extends BasePageClass {
-
-    @Test
-    public void test() {
-        System.out.println();
-    }
-
-    @After
-    public void teardown() {
-        // Закрой браузер
-        //driver.quit();
-    }
-}
