+++ src/test/java/site/nomoreparties/stellarburgers/LoginTest.java
@@ -1,0 +1,2 @@
+package site.nomoreparties.stellarburgers;public class LoginTest {
+}
